+++ packages/core-ts-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [2.3.8] (2025-11-08)
+
+### 🏗️ Changed
+
+- Updated dependencies for compatibility and security improvements
+
 ## 🔖 [2.3.7] (2025-11-07)
 
 ### 🏗️ Changed
@@ -106,6 +112,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Link References -->
 
+[2.3.8]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.7...core-ts-lib-v2.3.8
 [2.3.7]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.6...core-ts-lib-v2.3.7
 [2.3.6]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.5...core-ts-lib-v2.3.6
 [2.3.5]: https://github.com/aneuhold/ts-libs/compare/core-ts-lib-v2.3.4...core-ts-lib-v2.3.5